Fixes for warnings when building with MSVC
Most warnings are innocuous (wrong type-specifier for forward
declarations, conversion of literals into unsigned integers, warnings
about methods optimized out), but I believe actual bugs were revealed in
vartypedelegate.cpp and combat.cpp.

+++ apps/openmw/mwgui/bookpage.cpp
@@ -767,7 +767,7 @@ public:
 
     PageDisplay ()
     {
-        mPage = -1;
+        mPage = static_cast<size_t>(-1);
         mViewTop = 0;
         mViewBottom = 0;
         mFocusItem = NULL;
@@ -917,7 +917,7 @@ public:
             else
             {
                 mBook.reset ();
-                mPage = -1;
+                mPage = static_cast<size_t>(-1);
                 mViewTop = 0;
                 mViewBottom = 0;
             }
